refactor: god-file 拆分 — main.go 迁移逻辑提取到 migrations.go (#157)
cmd/server/main.go (1190 行) 拆分：~850 行数据库迁移函数（runMigrations +
addDevicesGeneratedColumns + extendDevicesTypeCheck + backfillDeviceUUIDs +
backfillSatelliteUUIDs + backfillOfflineSince + mergeDuplicateMACDevices +
applyUniqueIndexMigrations + applyIdentityIndexMigrations + extendScanRunStatusCheck）
提取到 cmd/server/migrations.go。

main.go 1190→337 行 (-72%)，保留 main() 入口 + initLogger + seedAdminUser +
listenAndServeWithRetry 等启动/wiring 逻辑。纯代码移动，无逻辑变更。

routes.go 拆分留后续（NewRouter 本身是 composition root，逻辑内聚）。
go test -race ./... 全绿。
